@@ -119,15 +119,15 @@
     "test": "node ./node_modules/vscode/bin/test"
   },
   "devDependencies": {
-    "typescript": "^3.1.4",
+    "typescript": "^4.1.3",
     "vscode": "^1.1.30",
     "mocha": "^2.3.3",
-    "@types/node": "^6.0.40",
+    "@types/node": "^10",
     "@types/mocha": "^2.2.32",
     "@types/tmp": "^0.0.33"
   },
   "dependencies": {
-    "viz.js": "^2.0.0",
+    "@hpcc-js/wasm": "^1.0.0",
     "opn": "^6.0.0",
     "tmp": "^0.0.33"
   }
